Implementation of Asset Cache Server
Currently, cache regeneration takes a long time, and is multiplied by every team member that needs to regen their cache when new assets are added.

The Cache Server is an attempt to solve that problem, by storing cached asset files in some central repository. If the server is enabled, when the AssetDatabase finds Assets in need of importing, it will first check with the remote server to see if it can pass back copies of cache files that some other user has previously generated.

If not, the local user is the first to generate those cache files, they will do so locally and then pass up the resulting files to the cache server for other users to have access to. With a decent network connection, this will hopefully be quicker than regenerating, for example, large DDS files.

To facilitate this an additional .md5 file is generated per asset in the cache, and used to later validate if a cache update is necessary or to request the appropriate pregenerated cached files from the Asset Cache Server. The server retains all cache files up until a specified maximum cache file is reached. At that point when a new file is pushed to the server, files are deleted using and oldest-first check until enough space is available for the file being pushed.

The Asset Cache Server is used by running AtomicTool with the appropriate parameters:
AtomicTool.exe cacheserver path/to/cach [port(default=20)] [space limit in MB(default=1024)]

To have the editor use the cache server, add a json file called AssetCache.json to the project's Settings directory:
{
	"useServer" : true|false,
	"serverIP" : "x.x.x.x",
	"serverPort" : number
}