Adding luma script bindings.
Not sure this is the right place for these script bindings, as it's outside our submodule folder and puts the burden on coders to remember to update the script bindings when the source changes, and then commit those changes to the AGE_Luma repo instead of our submodule, which is not desirable, I think. I suspect we can move these into our submodule then use cmake to get them where they need to be, but I'm not super proficient with it.

 - Matt to review.