(In conjunction with SHA f69b1fd0 on the LumaAtomicIP submodule)
Putting classes from 2 different namespaces in the js modules proved problem, it wasn't recognizing the one namespace which resulted in class inheritance and constructors being missing.

To solve that, split the luma project into two. The source files for both exist in the LumaAtomicIP submodule, the cmake file builds 2 projects out of that source and sets up dependencies etc.